+++ src/components/Wrappers/withUserRoles.js
@@ -53,6 +53,16 @@ const withUserRoles = (WrappedComponent) => (props) => {
     await Promise.allSettled(requests);
   };
 
+  const updateKeycloakUser = async (userId, data) => {
+    try {
+      await api.put(`users-keycloak/users/${userId}`, {
+        json: { ...data }
+      });
+    } catch (error) {
+      sendErrorCallout(error);
+    }
+  };
+
   const checkUserInKeycloak = async () => {
     try {
       await api.get(`users-keycloak/auth-users/${userId}`);
@@ -70,22 +80,26 @@ const withUserRoles = (WrappedComponent) => (props) => {
     await createKeycloakUser(userId);
   };
 
-  const handleKeycloakUserExists = async (onFinish) => {
-    await updateUserRoles(assignedRoleIds);
+  const handleKeycloakUserExists = async (onFinish, data) => {
+    await updateKeycloakUser(userId, data);
+
+    if (!isEqual(assignedRoleIds, initialAssignedRoleIds)) {
+      await updateUserRoles(assignedRoleIds);
+    }
     onFinish();
   };
 
-  const checkAndHandleKeycloakAuthUser = async (onFinish) => {
-    if (isEqual(assignedRoleIds, initialAssignedRoleIds)) {
-      onFinish();
-      return;
-    }
+  const checkAndHandleKeycloakAuthUser = async (onFinish, data, mutator) => {
     const userKeycloakStatus = await checkUserInKeycloak();
     switch (userKeycloakStatus) {
       case KEYCLOAK_USER_EXISTANCE.exist:
-        await handleKeycloakUserExists(onFinish);
+        // Only save changes to mod-users-keycloak.
+        await handleKeycloakUserExists(onFinish, data);
         break;
       case KEYCLOAK_USER_EXISTANCE.nonExist:
+        // First, save changes to mod-users.
+        // If user decides to create a Keycloak user, then changes will be copied over from mod-users to mod-users-keycloak.
+        await mutator.selUser.PUT(data);
         setIsCreateKeycloakUserConfirmationOpen(true);
         break;
       default: