+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## [12.1.3] (https://github.com/folio-org/ui-users/tree/v12.1.3) (2025-04-24)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v12.1.2...v12.1.3)
+* Ensure keycloak user updates are sent to `mod-users-keycloak` and non-keycloak user updates are sent to `mod-users`. Refs UIU-3360.
 * Add missed subPermission `ui-users.roles.view` to `ui-users.roles.manage`. Refs UIU-3363.
 
 ## [12.1.2] (https://github.com/folio-org/ui-users/tree/v12.1.2) (2025-04-10)